pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coffee-pipeline"
version = "0.1.0"
description = "AI Research Pipeline for Ba Tê và Cà Phê blog"
requires-python = ">=3.12"
dependencies = [
    "langgraph==1.1.3",
    "boto3>=1.38.0",
    "arxiv>=2.1.3",
    "crawl4ai>=0.5.0",
    "yt-dlp>=2025.1.1",
    "youtube-transcript-api>=0.6.3",
    "httpx>=0.28.0",
    "python-slugify[unidecode]>=8.0.4",
    "click>=8.1.8",
    "python-dotenv>=1.1.0",
    "openai>=1.0.0",
    "duckduckgo-search>=7.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.0.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "hypothesis>=6.100.0",
]

[project.scripts]
coffee-research = "coffee_pipeline.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/coffee_pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
